requirements.txt
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8
chromadb==0.5.0
openai==1.31.1
python-dotenv==1.0.1
tiktoken==0.7.0
